moving getUserByUsername to user file and also adding getUserById
